rtic_time::half_period_counter

Trait TimerOps

pub trait TimerOps: Copy {
    const MAX: Self;
    const ONE: Self;

    // Required methods
    fn xor(self, other: Self) -> Self;
    fn and(self, other: Self) -> Self;
    fn add(self, other: Self) -> Self;
    fn left_shift(self, amount: u32) -> Self;
}

Operations a type has to support in order to be used as the return value of calculate_now.

Required Associated Constants

const MAX: Self

All bits set to 1.

const ONE: Self

The lowest bit set to 1.

Required Methods

fn xor(self, other: Self) -> Self

The ^ operation.

fn and(self, other: Self) -> Self

The & operation.

fn add(self, other: Self) -> Self

The + operation.

fn left_shift(self, amount: u32) -> Self

The << operation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.